+++ assets/javascripts/kitten/components/links/link-list.js
@@ -19,6 +19,7 @@ export class LinkList extends Component {
       item,
       href,
       active,
+      title,
     } = element
 
     const { color } = this.props
@@ -40,6 +41,7 @@ export class LinkList extends Component {
         <a
           href={ href }
           className={ linkListClassName }
+          title={ title }
         >
           { item }
         </a>
@@ -80,4 +82,5 @@ LinkList.defaultProps = {
   margin: true,
   items: [], // Eg: [{ key: …, item: …, href: … }]
   color: 'dark',
+  title: '',
 }