+++ assets/javascripts/kitten/karl/links/link-list.js
@@ -3,10 +3,10 @@ import { LinkList } from 'kitten/components/links/link-list'
 
 export const KarlLinkList = () => {
   const linkListItems = [
-    { key: 'key1', item: 'Litterae', href: '#' },
-    { key: 'key2', item: 'Notae', href: '#' },
-    { key: 'key3', item: 'Bibliographia', href: '#' },
-    { key: 'key4', item: 'Nexus', href: '#' },
+    { key: 'key1', item: 'Litterae', href: '#', title: 'Litterae' },
+    { key: 'key2', item: 'Notae', href: '#', title: 'Notae' },
+    { key: 'key3', item: 'Bibliographia', href: '#', title: 'Bibliographia' },
+    { key: 'key4', item: 'Nexus', href: '#', title: 'Nexus' },
   ]
 
   return (
@@ -18,10 +18,16 @@ export const KarlLinkList = () => {
 
 export const KarlLinkListSelectedItem = () => {
   const linkListSelectedItem = [
-    { key: 'key1', item: 'Litterae', href: '#' },
-    { key: 'key2', item: 'Notae', href: '#' },
-    { key: 'key3', item: 'Bibliographia', href: '#', active: true },
-    { key: 'key4', item: 'Nexus', href: '#' },
+    { key: 'key1', item: 'Litterae', href: '#', title: 'Litterae' },
+    { key: 'key2', item: 'Notae', href: '#', title: 'Notae' },
+    {
+      key: 'key3',
+      item: 'Bibliographia',
+      href: '#',
+      title: 'Bibliographia',
+      active: true,
+    },
+    { key: 'key4', item: 'Nexus', href: '#', title: 'Nexus' },
   ]
 
   return (
@@ -33,10 +39,10 @@ export const KarlLinkListSelectedItem = () => {
 
 export const KarlLinkListDark = () => {
   const linkListItems = [
-    { key: 'key1', item: 'Litterae', href: '#' },
-    { key: 'key2', item: 'Notae', href: '#', active: true },
-    { key: 'key3', item: 'Bibliographia', href: '#' },
-    { key: 'key4', item: 'Nexus', href: '#' },
+    { key: 'key1', item: 'Litterae', href: '#', title: 'Litterae' },
+    { key: 'key2', item: 'Notae', href: '#', title: 'Notae', active: true },
+    { key: 'key3', item: 'Bibliographia', href: '#', title: 'Bibliographia' },
+    { key: 'key4', item: 'Nexus', href: '#', title: 'Nexus' },
   ]
 
   return (