+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add A11Y on `LinkList`, `InstagramIcon`, `FacebookIcon`, `TwitterIcon`.
+- Feature: Add `placeholder` prop to `TextInputWithButton` component.
 - Feature: Add `k-u-color-background1` utility.
 
 ## [19.0.0] - 2018-01-16